server/requirements-serve.txt
# Web server extra requirements
Flask-HTTPAuth
python-dotenv
